AG-13306 Move Zoom/ZoomDOMProxy testFindTarget into findTarget.ts
