+++ packages/ag-charts-enterprise/src/features/zoom/zoomDOMProxy.ts
@@ -69,18 +69,6 @@ export class ZoomDOMProxy {
         }
     }
 
-    testFindTarget(canvasX: number, canvasY: number): _ModuleSupport.MockEvent | undefined {
-        for (const axis of this.axes) {
-            const bbox = axis.div.getBounds();
-            if (!axis.div.isHidden() && BBoxValues.containsPoint(bbox, canvasX, canvasY)) {
-                const offsetX = canvasX - bbox.x;
-                const offsetY = canvasY - bbox.y;
-                return { target: axis.div.getElement(), offsetX, offsetY };
-            }
-        }
-        return undefined;
-    }
-
     private diffAxisIds(axesCtx: _ModuleSupport.AxisContext[]) {
         const myIds = this.axes.map((entry) => entry.axisId);
         const ctxIds = axesCtx.map((ctx) => ctx.axisId);
